+++ integration/fuse/src/main/java/alluxio/cli/FuseCommand.java
@@ -9,10 +9,9 @@
  * See the NOTICE file distributed with this work for information regarding copyright ownership.
  */
 
-package alluxio.fuse.cli;
+package alluxio.cli;
 
 import alluxio.AlluxioURI;
-import alluxio.cli.Command;
 import alluxio.client.file.URIStatus;
 import alluxio.exception.status.InvalidArgumentException;
